+++ frontend/src/components/main/main_page.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import LoginForm from '../session/login_form';
+import LoginForm from '../session/login_form_container';
 
 class MainPage extends React.Component {
 
   render() {
     return (
       <div>
-        <LoginForm/>
+        <h1>Main</h1>
         <footer>
         </footer>
       </div>
