+++ frontend/src/util/route_util.js
@@ -7,7 +7,7 @@ const Auth = ({ component: Component, path, loggedIn, exact }) => (
     !loggedIn ? (
       <Component {...props} />
     ) : (
-      <Redirect to="/tweets" />
+      <Redirect to="/" />
     )
   )} />
 );
