+++ models/Clinician.js
@@ -2,11 +2,11 @@ const BaseUser = require('./BaseUser');
 const mongoose = require('mongoose');  
 const Schema = mongoose.Schema;
 
-let ClinitianSchema = new BaseUser.BaseUserSchema({
+let ClinicianSchema = new BaseUser.BaseUserSchema({
     patients: [{ type: Schema.Types.ObjectId, ref: 'Patient' }],
 })
 
-BaseUser.User.discriminator('Clinitian', ClinitianSchema);
+BaseUser.User.discriminator('Clinician', ClinicianSchema);
 
-const Clinitian = mongoose.model('Clinitian');
-module.exports = Clinitian;
+const Clinician = mongoose.model('Clinician');
+module.exports = Clinician;
