+++ models/Patient.js
@@ -3,7 +3,7 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 let PatientSchema = new BaseUser.BaseUserSchema({
-    clinician: { type: Schema.Types.ObjectId, ref: 'Clinitian' },
+    clinician: { type: Schema.Types.ObjectId, ref: 'Clinician' },
     // exercises: [{ type: Schema.Types.ObjectId, ref: 'Exercise'}],
 })
